Conflicts:
	SurgSim/DataStructures/UnitTests/CMakeLists.txt
	SurgSim/DataStructures/UnitTests/TriangleMeshTest.cpp

+++ SurgSim/Devices/Novint/NovintScaffold.cpp
@@ -605,7 +605,7 @@ bool NovintScaffold::updateDevice(DeviceData* info)
 	fatalError = checkForFatalError(fatalError, "hdlGripGetAttributev(HDL_GRIP_POSITION)");
 
 	info->buttonStates.fill(false);
-	hdlGripGetAttributesb(HDL_GRIP_BUTTON, info->buttonStates.size(), info->buttonStates.data());
+	hdlGripGetAttributesb(HDL_GRIP_BUTTON, static_cast<int>(info->buttonStates.size()), info->buttonStates.data());
 	fatalError = checkForFatalError(fatalError, "hdlGripGetAttributesb(HDL_GRIP_BUTTON)");
 
 	// Get the additional 7DoF data if available.
